Fix signed-integer-overflow UB in floatEquals()'s ULP subtraction
After the previous commit's sign-bit remap fix, A.i32/B.i32 can each
land anywhere across nearly the full int32_t range (an extreme
positive float remaps to roughly +2^31-1, an extreme negative one to
roughly -(2^31-1)). Computing their difference as plain int32_t
subtraction can then itself overflow int32_t -- genuine signed-integer-
overflow UB, and easily reachable in practice: floatEquals(100.0f,
-100.0f, ...) already triggers it, not just astronomical values.

Widen to int64_t (comfortably holds any int32_t difference) before
subtracting. Added floatEquals_oppositeSignOverflow covering both a
moderate-magnitude and an extreme opposite-sign pair, plus a sanity
check that small values straddling zero (using denorm_min(), exactly
1 ULP from 0.0f, rather than a decimal-looking-small value that's
actually tens of thousands of ULPs away) still compare almost-equal.

Verified clean under GCC Debug ASan+UBSan (no signed-overflow report,
329 tests / 83574 checks pass) and under plain GCC/clang builds +
ctest.

+++ testsuite/TestSuiteMisc.h
@@ -81,8 +81,16 @@ inline bool floatEquals(float Ain, float Bin, unsigned int maxUlps)
     // Make B.i32 lexicographically ordered as a twos-complement int
     if (B.i32 < 0)
         B.i32 = 0x80000000 - B.i32;
-    unsigned int intDiff = SbAbs(A.i32 - B.i32);
-    if (intDiff <= maxUlps)
+    // A.i32/B.i32 can each land anywhere across nearly the full int32_t
+    // range after the remap above (an extreme positive float remaps to
+    // roughly +2^31-1, an extreme negative one to roughly -(2^31-1)), so
+    // a plain "A.i32 - B.i32" can itself overflow int32_t -- genuine
+    // signed-integer-overflow UB, not just a theoretical concern: e.g.
+    // floatEquals(100.0f, -100.0f, ...) already hits it. Widen to
+    // int64_t, which comfortably holds any int32_t difference, before
+    // subtracting.
+    int64_t intDiff = SbAbs(static_cast<int64_t>(A.i32) - static_cast<int64_t>(B.i32));
+    if (intDiff <= static_cast<int64_t>(maxUlps))
         return true;
     return false;
 }
